Upgrade Qt5 (#71)
* Add a patch for ZooKeeper

The ZooKeeper project no longer builds with modern Java. I
added a patch to update the source and target properties in
the build.xml file.

* Partial fix of a QWT build error

QWT can sometimes fail to find qmake. I updated QWT's
dependency from the QT5 build target to the QT5 install
target. This doesn't completely fix the issue, but it helps.

Issue #65

* Update Qt5

I updated Qt from 5.11 to 5.12. Qt5 removed one of the
configure options, so I had to take that out of our project
CMake file. Qt5 also removed a text file that we copied to
the installation, so I had to remove that copy operation.

Closes #70